requirements.txt
ultralytics
google-cloud-storage # ローカルテスト時はコメントアウト
PyYAML